requirements.txt
PyYAML==6.0.1
requests==2.31.0
GitPython==3.1.43
